Local storage may work but Edge doesnt allow it

+++ app/Controllers/ListsController.js
@@ -2,33 +2,38 @@ import { ProxyState } from "../AppState.js";
 import { listsService } from "../Services/ListsService.js";
 import { loadState, saveState } from "../Utils/LocalStorage.js";
 import TasksController from "./TasksController.js";
-let tasksController = new TasksController()
+
 
 
 function _draw(){
   let template = ''
   let lists = ProxyState.lists
-  lists.forEach(list => {
-    //Getting total and all checked tasks: can make these functions later
-    let totTask = ProxyState.tasks.filter(task=> task.listID == list.listID)
-    let finTask = 0
-    if(totTask.length == 0){
-      template += list.getTemplate(0, 0)
-    } else {
-      totTask.forEach(task=> {
-        if(task.isSelected){
-          finTask++
-      }})
-    template += list.getTemplate(finTask, totTask.length)
-    }
-  document.getElementById('lists').innerHTML = template
-  })
+  if(lists.length == 0){
+    document.getElementById('lists').innerHTML = template
+  } else {
+    lists.forEach(list => {
+      //Getting total and all checked tasks: can make these functions later
+      let totTask = ProxyState.tasks.filter(task => task.listID == list.listID)
+      let finTask = 0
+      if (totTask.length == 0) {
+        template += list.getTemplate(0, 0)
+      } else {
+        totTask.forEach(task => {
+          if (task.isSelected) {
+            finTask++
+          }
+        })
+        template += list.getTemplate(finTask, totTask.length)
+      }
+      document.getElementById('lists').innerHTML = template
+    })
+  }
 }
 
 export default class ListsController{
 
   constructor(){
-    ProxyState.on['lists', _draw]
+    ProxyState.on['lists', _draw()]
     ProxyState.on['tasks', _draw]
     ProxyState.on['lists', saveState]
     ProxyState.on['tasks', saveState]
@@ -42,10 +47,14 @@ export default class ListsController{
   }
 
   deleteList(listId){
-    listsService.deleteList(listId)
-    tasksController.deleteTasksInList(listId)
-    _draw()
+    if(window.confirm("Delete List?")){
+      let tasksController = new TasksController()
+      listsService.deleteList(listId)
+      tasksController.deleteTasksInList(listId)
+      _draw()
+    }
   }
+
 
   addList(){
     event.preventDefault()
